shark_sentinel/requirements.txt
Django
djangorestframework
rasterio
numpy
matplotlib
